refactor(websocketx): 将 Response 结构体的 Code 字段类型从 uint32 改为 int32
- 修改了 Response 结构体中 Code 字段的类型，以适应 code.Code 的 int32 类型
- 在返回 Response 时，对 code.Code进行了类型转换，确保数据类型一致

+++ infrastructure/pkg/websocketx/process.go
@@ -45,7 +45,7 @@ func (h *Head) String() (headStr string) {
 
 // Response 响应数据体
 type Response struct {
-	Code uint32      `json:"code"`
+	Code int32       `json:"code"`
 	Msg  string      `json:"msg"`
 	Data interface{} `json:"data"` // 数据 json
 }
@@ -66,7 +66,7 @@ func NewJsonResponse(code *e.StatusCode, data []byte) *Response {
 	if err != nil {
 		logx.Errorf("json.Unmarshal error: %v", err)
 	}
-	return &Response{Code: uint32(code.Code), Msg: code.Message, Data: result}
+	return &Response{Code: int32(uint32(code.Code)), Msg: code.Message, Data: result}
 }
 
 // NewProtoMessageResponse 设置返回消息
